Create shareable ZonePlotter class
Refactor to reuse common plotting layouts and saving to HTML.
