+++ notebooks/nb_shared.py
@@ -1,7 +1,7 @@
 import xarray as xr
 
 from swed_17.nb_helpers import start_cluster
-from swed_17 import Snow17SWE, ZoneCompare, ZoneDB
+from swed_17 import Snow17SWE, ZoneCompare, ZoneDB, ZonePlotter
 from swed_17.peak_swe import peak_swe_for_pd, plot_peak_swe_pd
 from swed_17.swann_helpers import swann_data_for_zone
