+++ src/swed_17/zone_compare.py
@@ -1,9 +1,6 @@
 import holoviews as hv
 import pandas as pd
-import panel as pn
 import xarray as xr
-
-from bokeh.resources import INLINE
 
 
 class ZoneCompare:
@@ -87,7 +84,7 @@ class ZoneCompare:
             self.swann.hvplot('time', label='SWANN')
         ])
 
-    def plot_all(self) -> hv.Layout:
+    def plot(self) -> hv.Layout:
         """
         Plot the time series and correlation side-by-side
 
@@ -98,42 +95,3 @@ class ZoneCompare:
         return hv.Layout(
             self.time_series_plot() + self.correlation_plot()
         ).cols(2)
-
-    @classmethod
-    def show_all(cls, collection: list) -> hv.Layout:
-        """
-        Show collection of zone compare instances in one layout with two columns.
-
-        For use in notebooks.
-
-        Parameters
-        ----------
-        collection : list
-            Collection of ZoneCompare instances
-
-        Returns
-        -------
-        hv.Layout
-            Holoviews layout with all plots
-        """
-        plots = [zone_compare.plot_all() for zone_compare in collection]
-        return hv.Layout(plots).opts(shared_axes=False).cols(2)
-
-    @classmethod
-    def save_html(cls, file_path: str, collection: list) -> None:
-        """
-        Save plots as interactive HTML page
-
-        Parameters
-        ----------
-        file_path : str
-            Path to save files to
-        collection : list
-            Collection of ZoneCompare instances
-        """
-
-        pn.pane.HoloViews(
-            cls.show_all(collection)
-        ).save(
-            file_path, embed=True, resources=INLINE
-        )
