+++ src/swed_17/zone_plotter.py
@@ -1,0 +1,47 @@
+import holoviews as hv
+import panel as pn
+
+from bokeh.resources import INLINE
+
+
+class ZonePlotter:
+    """
+    Utility class for plotting and export to HTML pages
+    """
+    @classmethod
+    def show_all(cls, collection: list) -> hv.Layout:
+        """
+        Show given collection in one layout with two columns.
+
+        For use in notebooks.
+
+        Parameters
+        ----------
+        collection : list
+            Collection of Holoview plots
+
+        Returns
+        -------
+        hv.Layout
+            Holoviews layout with all plots in two columns
+        """
+        return hv.Layout(collection).opts(shared_axes=False).cols(2)
+
+    @classmethod
+    def save_html(cls, file_path: str, collection: list) -> None:
+        """
+        Save plots as interactive HTML page
+
+        Parameters
+        ----------
+        file_path : str
+            Path to save files to
+        collection : list
+            Collection of ZoneCompare instances
+        """
+
+        pn.pane.HoloViews(
+            cls.show_all(collection)
+        ).save(
+            file_path, embed=True, resources=INLINE
+        )
